WIP refactored 'actions' to an enum
